requirements.txt
aiogram>=3,<4
matplotlib
python-dotenv
